refactor: split picture item stateless to avoid circular dependency
