+++ apps/renderer/src/modules/entry-column/item.tsx
@@ -11,14 +11,10 @@ import type { FlatEntryModel } from "~/store/entry"
 import { useEntry } from "~/store/entry/hooks"
 
 import { ReactVirtuosoItemPlaceholder } from "../../components/ui/placeholder"
-import {
-  getItemComponentByView,
-  getSkeletonItemComponentByView,
-  getStatelessItemComponentByView,
-} from "./Items"
+import { getItemComponentByView, getSkeletonItemComponentByView } from "./Items"
 import { EntryItemWrapper } from "./layouts/EntryItemWrapper"
 import { girdClassNames } from "./styles"
-import type { EntryItemStatelessProps, EntryListItemFC } from "./types"
+import type { EntryListItemFC } from "./types"
 
 interface EntryItemProps {
   entryId: string
@@ -55,12 +51,6 @@ export const EntryItem: FC<EntryItemProps> = memo(({ entryId, view }) => {
 
   if (!entry) return <ReactVirtuosoItemPlaceholder />
   return <EntryItemImpl entry={entry} view={view} />
-})
-
-export const EntryItemStateless: FC<EntryItemStatelessProps> = memo((props) => {
-  const Item = getStatelessItemComponentByView(props.view as FeedViewType)
-
-  return <Item {...props} />
 })
 
 const LoadingCircleFallback = (
